Add files via upload
A bunch of jupyter notebooks for loading, reshaping/cleaning, and imputing the data, an edited version of the artist similarity graph script, and a weighted cosine similarity function for the end. 

Notes on use if you decide to use these yourself (I don't really recommend it, it takes a while, but if need be): 

If you're checking out the similarity function, use this notebook first. This will get you the non-nested columns and the pitches. To get the timbres, I'd recommend using "nested_column_commandline" and, as its name suggests, running it from the command line. Also, I removed columns that were more than 95% nan before doing the imputing, and split my work over several notebooks. I did all of this because of the fact that this is all pretty memory-intensive, and if I didn't chunk things up one at a time, Python couldn't allocate enough memory for a lot of this.

So once you've done that, run "nested_column_commandline". After that, run "do_artist_terms". Then run "prep_and_run_cosine_similarity." Those last two will give you PCAs (and in case of timbre/pitch, run some PCAs to make things a lot faster). "prep_and_run_cosine_similarity" also, as the name suggests, contains the similarity function.